Refactor memory functions to use size_t instead of uint16_t for better compatibility and consistency

+++ src/include/memory.h
@@ -5,9 +5,9 @@
 
 #define kfree(a) _kfree((void**)&a)
 
-void* memcpy(void* dst, const void* src, uint16_t num);
-void* memset(void* ptr, int value, uint16_t num);
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
+void* memcpy(void* dst, const void* src, size_t num);
+void* memset(void* ptr, int value, size_t num);
+int memcmp(const void* ptr1, const void* ptr2, size_t num);
 
 void* kmalloc(size_t size);
 void _kfree(void **ptr);
